student repository

+++ src/main/java/ir/mapsa/secondspringproject/tutorials1/StudentRepository.java
@@ -7,16 +7,23 @@ import java.util.ArrayList;
 import java.util.List;
 
 @Service
-public class StudentRepository implements BaseStudentRepository {
+public class StudentRepository implements BaseRepository<Student> {
     @Override
     public void add(Student student) throws Exception {
+        executeUpdateQuery("insert into second_student(name, family, passed_course, student_id, national_code) values (?,?,?,?,?)", student);
+    }
+
+    private void executeUpdateQuery(String query,Student student) throws Exception {
         try (Connection connection = getConnection()) {
-            try (PreparedStatement statement = connection.prepareStatement("insert into second_student(name, family, passed_course, student_id, national_code) values (?,?,?,?,?)")) {
+            try (PreparedStatement statement = connection.prepareStatement(query)) {
                 statement.setString(1, student.getName());
                 statement.setString(2, student.getFamily());
                 statement.setInt(3, student.getPassedCourse());
                 statement.setString(4, student.getStudentId());
                 statement.setString(5, student.getNationalCode());
+                if (student.getId() != null) {
+                    statement.setLong(6, student.getId());
+                }
                 statement.executeUpdate();
             }
         }
@@ -24,22 +31,15 @@ public class StudentRepository implements BaseStudentRepository {
 
     @Override
     public void update(Student student) throws Exception {
-        try (Connection connection = getConnection()) {
-            try (PreparedStatement statement = connection.prepareStatement("update second_student set name=?, family=?, passed_course=?, student_id=?, national_code=?\n" +
-                    "where id=?")) {
-                statement.setString(1, student.getName());
-                statement.setString(2, student.getFamily());
-                statement.setInt(3, student.getPassedCourse());
-                statement.setString(4, student.getStudentId());
-                statement.setString(5, student.getNationalCode());
-                statement.setLong(6, student.getId());
-                statement.executeUpdate();
-            }
-        }
+        executeUpdateQuery("update second_student set name=?, family=?, passed_course=?, student_id=?, national_code=?\n" +
+                "where id=?",student);
     }
 
     @Override
     public void removeById(Long id) throws Exception {
+        if (this.findById(id) == null) {
+            throw new IllegalArgumentException();
+        }
         try (Connection connection = getConnection()) {
             try (PreparedStatement statement = connection.prepareStatement("delete from second_student where id=?")) {
                 statement.setLong(1, id);
@@ -56,21 +56,26 @@ public class StudentRepository implements BaseStudentRepository {
                 statement.setLong(1, id);
                 try (ResultSet resultSet = statement.executeQuery()) {
                     if (resultSet.next()) {
-                        Student student = new Student();
-                        student.setStudentId(resultSet.getString("student_id"));
-                        student.setFamily(resultSet.getString("family"));
-                        student.setName(resultSet.getString("name"));
-                        student.setNationalCode(resultSet.getString("national_code"));
-                        student.setPassedCourse(resultSet.getInt("passed_course"));
-                        student.setId(id);
+                        Student student = getStudent(resultSet);
                         return student;
 
-                    }else{
+                    } else {
                         return null;
                     }
                 }
             }
         }
+    }
+
+    private static Student getStudent( ResultSet resultSet) throws SQLException {
+        Student student = new Student();
+        student.setStudentId(resultSet.getString("student_id"));
+        student.setFamily(resultSet.getString("family"));
+        student.setName(resultSet.getString("name"));
+        student.setNationalCode(resultSet.getString("national_code"));
+        student.setPassedCourse(resultSet.getInt("passed_course"));
+        student.setId(resultSet.getLong("id"));
+        return student;
     }
 
     private Connection getConnection() throws ClassNotFoundException, SQLException {
@@ -79,29 +84,31 @@ public class StudentRepository implements BaseStudentRepository {
     }
 
     @Override
-    public List<Student> getAll() throws Exception{
+    public List<Student> getAll() throws Exception {
         try (Connection connection = getConnection()) {
             try (PreparedStatement statement = connection.prepareStatement("select * from second_student")) {
-
-
                 try (ResultSet resultSet = statement.executeQuery()) {
                     List<Student> result = new ArrayList<>();
-                    while(resultSet.next()) {
-                        Student student = new Student();
-                        student.setStudentId(resultSet.getString("student_id"));
-                        student.setFamily(resultSet.getString("family"));
-                        student.setName(resultSet.getString("name"));
-                        student.setNationalCode(resultSet.getString("national_code"));
-                        student.setPassedCourse(resultSet.getInt("passed_course"));
-                        student.setId(resultSet.getLong("id"));
+                    while (resultSet.next()) {
+                        Student student = getStudent(resultSet);
                         result.add(student);
                     }
                     return result;
                 }
             }
         }
+    }
 
+    public List<Student> findByExample(Student student) {
+        String query = "select * from second_student where 1=1 ";
+        if (student.getId() != null) {
+            query += " AND id = "+student.getId();
+        }
 
-
+        if (student.getStudentId() != null) {
+            query += " AND student_id = '"+ student.getStudentId()+"'";
+        }
+        //todo implement this method
+        return null;
     }
 }
